Fixed merge issue and moved acq time cycle function out of rtc query
